+++ jdbc/src/main/java/nextstep/jdbc/JdbcTemplate.java
@@ -10,6 +10,9 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 import javax.sql.DataSource;
+import nextstep.jdbc.exception.EmptyResultDataAccessException;
+import nextstep.jdbc.exception.IncorrectResultSizeDataAccessException;
+import nextstep.jdbc.exception.JdbcConnectionException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
