+++ jdbc/src/main/java/nextstep/jdbc/exception/IncorrectResultSizeDataAccessException.java
@@ -1,4 +1,4 @@
-package nextstep.jdbc;
+package nextstep.jdbc.exception;
 
 public class IncorrectResultSizeDataAccessException extends RuntimeException {
