+++ jdbc/src/main/java/nextstep/jdbc/exception/JdbcConnectionException.java
@@ -1,4 +1,4 @@
-package nextstep.jdbc;
+package nextstep.jdbc.exception;
 
 public class JdbcConnectionException extends RuntimeException {
